change visibility of mapCodec
make test of codec provider simpler
